+++ tacacs-plus-protocol/src/lib.rs
@@ -28,7 +28,7 @@ mod fields;
 pub use fields::*;
 
 mod text;
-pub use text::FieldText;
+pub use text::{FieldText, InvalidText};
 
 #[cfg(feature = "std")]
 mod owned;
@@ -148,13 +148,16 @@ impl fmt::Display for DeserializeError {
 #[cfg(feature = "std")]
 mod error_impls {
     use std::error::Error;
+    use std::fmt;
 
+    use super::text::InvalidText;
     use super::{DeserializeError, InvalidArgument, SerializeError};
 
     impl Error for DeserializeError {}
     impl Error for SerializeError {}
     impl Error for InvalidArgument {}
     impl Error for super::authentication::BadStart {}
+    impl<T> Error for InvalidText<T> where InvalidText<T>: fmt::Debug + fmt::Display {}
 }
 
 // suggestion from Rust API guidelines: https://rust-lang.github.io/api-guidelines/future-proofing.html#sealed-traits-protect-against-downstream-implementations-c-sealed