+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - CI tests are also run against [TACACS+ NG], an actively maintained TACACS+ server implementation (#30)
+- `InvalidText` type, for `FieldText` construction errors
 
 #### Changed
 
@@ -20,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ContextBuilder::new()` takes a `String` instead of an `&str`
 
 [TACACS+ NG]: https://projects.pro-bono-publico.de/event-driven-servers/doc/tac_plus-ng.html
+
+#### Changed
+
+- `FieldText`'s `TryFrom` & `FromStr` implementation error types were changed to `InvalidText`
 
 ### tacacs-plus-protocol